Refactored PouchDB/ConfigDB into single service
