+++ app/patients/edit/controller.js
@@ -5,8 +5,8 @@ import GenderList from 'hospitalrun/mixins/gender-list';
 import ReturnTo from 'hospitalrun/mixins/return-to';
 import UserSession from "hospitalrun/mixins/user-session";
 export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, UserSession, {
-    
-    canAddAppointment: function() {        
+
+    canAddAppointment: function() {
         return this.currentUserCan('add_appointment');
     }.property(),
 
@@ -16,61 +16,61 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
 
     canAddImaging: function() {
         return this.currentUserCan('add_imaging');
-    }.property(),    
+    }.property(),
 
-    canAddLab: function() {        
+    canAddLab: function() {
         return this.currentUserCan('add_lab');
-    }.property(),    
-    
-    canAddMedication: function() {        
+    }.property(),
+
+    canAddMedication: function() {
         return this.currentUserCan('add_medication');
     }.property(),
-    
+
     canAddPhoto: function() {
         var isFileSystemEnabled = this.get('isFileSystemEnabled');
         return (this.currentUserCan('add_photo') && isFileSystemEnabled);
     }.property(),
-    
-    canAddSocialWork: function() {        
+
+    canAddSocialWork: function() {
         return this.currentUserCan('add_socialwork');
     }.property(),
-    
-    canAddVisit: function() {        
+
+    canAddVisit: function() {
         return this.currentUserCan('add_visit');
     }.property(),
 
-    canDeleteAppointment: function() {        
+    canDeleteAppointment: function() {
         return this.currentUserCan('delete_appointment');
-    }.property(), 
-    
-    canDeleteContact: function() {        
+    }.property(),
+
+    canDeleteContact: function() {
         return this.currentUserCan('add_patient');
     }.property(),
-    
+
     canDeleteImaging: function() {
         return this.currentUserCan('delete_imaging');
-    }.property(),        
-    
-    canDeleteLab: function() {        
+    }.property(),
+
+    canDeleteLab: function() {
         return this.currentUserCan('delete_lab');
-    }.property(),        
-    
-    canDeleteMedication: function() {        
+    }.property(),
+
+    canDeleteMedication: function() {
         return this.currentUserCan('delete_medication');
     }.property(),
-    
-    canDeletePhoto: function() {        
+
+    canDeletePhoto: function() {
         return this.currentUserCan('delete_photo');
-    }.property(),    
-    
-    canDeleteSocialWork: function() {        
+    }.property(),
+
+    canDeleteSocialWork: function() {
         return this.currentUserCan('delete_socialwork');
     }.property(),
 
-    canDeleteVisit: function() {        
+    canDeleteVisit: function() {
         return this.currentUserCan('delete_visit');
     }.property(),
-    
+
     economicClassificationTypes: [
         'A',
         'B',
@@ -87,23 +87,23 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
         'Rent',
         'Shared'
     ],
-    
+
     patientTypes: [
         'Charity',
         'Private'
     ],
-    
-    philhealthTypes: [         
+
+    philhealthTypes: [
         'Employed: Government',
         'Employed: Non Paying Member/Lifetime',
         'Employed: OWWA/OFW',
         'Employed: Private',
         'Employed: Sponsored/Indigent',
-        'Self Employed'      
-    ],    
-    
+        'Self Employed'
+    ],
+
     filesystem: Ember.inject.service(),
-    pouchdb: Ember.inject.service(),
+    database: Ember.inject.service(),
     patientController: Ember.inject.controller('patients'),
 
     addressOptions: Ember.computed.alias('patientController.addressOptions'),
@@ -119,10 +119,10 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
     clinicList: Ember.computed.alias('patientController.clinicList'),
     countryList: Ember.computed.alias('patientController.countryList'),
     isFileSystemEnabled: Ember.computed.alias('filesystem.isFileSystemEnabled'),
-    
+
     pricingProfiles: Ember.computed.alias('patientController.pricingProfiles'),
     statusList: Ember.computed.alias('patientController.statusList'),
-    
+
     haveAdditionalContacts: function() {
         var additionalContacts = this.get('additionalContacts');
         return (!Ember.isEmpty(additionalContacts));
@@ -131,7 +131,7 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
     haveAddressOptions: function() {
         var addressOptions = this.get('addressOptions');
         return (!Ember.isEmpty(addressOptions));
-    }.property('addressOptions'),    
+    }.property('addressOptions'),
 
     lookupListsToUpdate: [{
         name: 'countryList',
@@ -146,25 +146,25 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
         property: 'status',
         id: 'patient_status_list'
     }],
-    
+
     patientImaging: function() {
         return this._getVisitCollection('imaging');
-    }.property('visits.@each.imaging'),    
-    
+    }.property('visits.@each.imaging'),
+
     patientLabs: function() {
         return this._getVisitCollection('labs');
     }.property('visits.@each.labs'),
-    
+
     patientMedications: function() {
         return this._getVisitCollection('medication');
     }.property('visits.@each.medication'),
-    
+
     patientProcedures: function() {
         return this._getVisitCollection('procedures');
     }.property('visits.@each.procedures'),
-    
+
     showExpenseTotal: true,
-    
+
     totalExpenses: function() {
         var expenses = this.get('expenses');
         if (!Ember.isEmpty(expenses)) {
@@ -172,13 +172,13 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                 if (!Ember.isEmpty(expense.cost)) {
                     return previousValue +  parseInt(expense.cost);
                 }
-            }, 0); 
+            }, 0);
             this.set('showExpenseTotal', true);
             return total;
         } else {
             this.set('showExpenseTotal', false);
         }
-    }.property('expenses'),    
+    }.property('expenses'),
 
     updateCapability: 'add_patient',
 
@@ -188,7 +188,7 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
             additionalContacts.addObject(newContact);
             this.set('additionalContacts', additionalContacts);
             this.send('update', true);
-            this.send('closeModal');            
+            this.send('closeModal');
         },
 
         /**
@@ -202,16 +202,16 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                 fileSystem = this.get('filesystem'),
                 photos = this.get('model.photos'),
                 newPatientPhoto = this.get('store').createRecord('photo', {
-                    patient: this.get('model'),                    
+                    patient: this.get('model'),
                     localFile: true,
                     caption: caption,
                     coverImage: coverImage,
                 });
             newPatientPhoto.save().then(function(savedPhotoRecord) {
-                var pouchDbId = this.get('pouchdb').getPouchId(savedPhotoRecord.get('id'),'photo');
+                var pouchDbId = this.get('database').getPouchId(savedPhotoRecord.get('id'),'photo');
                 fileSystem.addFile(photoFile, dirToSaveTo, pouchDbId).then(function(fileEntry) {
                     fileSystem.fileToDataURL(photoFile).then(function(photoDataUrl) {
-                        savedPhotoRecord = this.get('store').find('photo', savedPhotoRecord.get('id')).then(function(savedPhotoRecord) {                        
+                        savedPhotoRecord = this.get('store').find('photo', savedPhotoRecord.get('id')).then(function(savedPhotoRecord) {
                             var dataUrlParts = photoDataUrl.split(',');
                             savedPhotoRecord.setProperties({
                                 fileName: fileEntry.fullPath,
@@ -221,8 +221,8 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                                         content_type: photoFile.type,
                                         data: dataUrlParts[1]
                                     }
-                                }                    
-                            });                        
+                                }
+                            });
                             savedPhotoRecord.save().then(function(savedPhotoRecord) {
                                 photos.addObject(savedPhotoRecord);
                                 this.send('closeModal');
@@ -232,13 +232,13 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                 }.bind(this));
             }.bind(this));
         },
-        
+
         appointmentDeleted: function(deletedAppointment) {
-            var appointments = this.get('appointments');            
+            var appointments = this.get('appointments');
             appointments.removeObject(deletedAppointment);
             this.send('closeModal');
         },
-        
+
         deleteContact: function(model) {
             var contact = model.get('contactToDelete');
             var additionalContacts = this.get('additionalContacts');
@@ -273,12 +273,12 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                 var fileSystem = this.get('filesystem'),
                     isFileSystemEnabled = this.get('isFileSystemEnabled');
                 if (isFileSystemEnabled) {
-                    var pouchDbId = this.get('pouchdb').getPouchId(photoId,'photo');                       
+                    var pouchDbId = this.get('database').getPouchId(photoId,'photo');
                     fileSystem.deleteFile(filePath, pouchDbId);
                 }
             }.bind(this));
         },
-        
+
         editAppointment: function(appointment) {
             appointment.set('returnToPatient', true);
             this.transitionToRoute('appointments.edit', appointment);
@@ -291,8 +291,8 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                 });
                 this.transitionToRoute('imaging.edit', imaging);
             }
-        },        
-        
+        },
+
         editLab: function(lab) {
             if (lab.get('canEdit')) {
                 lab.setProperties({
@@ -300,64 +300,64 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                 });
                 this.transitionToRoute('labs.edit', lab);
             }
-        },        
-        
+        },
+
         editMedication: function(medication) {
             if (medication.get('canEdit')) {
                 medication.set('returnToPatient', true);
                 this.transitionToRoute('medication.edit', medication);
             }
-        },    
-        
-        editPhoto: function(photo) {        
+        },
+
+        editPhoto: function(photo) {
             this.send('openModal', 'patients.photo', photo);
         },
-        
+
         editProcedure: function(procedure) {
             this.transitionToRoute('procedures.edit', procedure);
         },
-        
+
         editVisit: function(visit) {
             this.transitionToRoute('visits.edit', visit);
         },
-        
-        newAppointment: function() {            
+
+        newAppointment: function() {
             this._addChildObject('appointments.edit');
         },
 
         newImaging: function() {
             this._addChildObject('imaging.edit');
         },
-        
+
         newLab: function() {
-            this._addChildObject('labs.edit');            
+            this._addChildObject('labs.edit');
         },
-        
+
         newMedication: function() {
             this._addChildObject('medication.edit');
         },
-        
+
         newVisit: function() {
             var patient = this.get('model'),
                 visits = this.get('visits');
             this.send('createNewVisit', patient, visits);
-        },     
+        },
 
         showAddContact: function() {
-            this.send('openModal', 'patients.add-contact', {});            
+            this.send('openModal', 'patients.add-contact', {});
         },
 
         showAddPhoto: function() {
             this.send('openModal', 'patients.photo', {
                 isNew: true
             });
-        },        
-        
+        },
+
         showDeleteAppointment: function(appointment) {
             appointment.set('deleteFromPatient', true);
             this.send('openModal', 'appointments.delete', appointment);
         },
-        
+
         showDeleteContact: function(contact) {
             this.send('openModal', 'dialog', Ember.Object.create({
                 confirmAction: 'deleteContact',
@@ -366,9 +366,9 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                 contactToDelete: contact,
                 updateButtonAction: 'confirm',
                 updateButtonText: 'Ok'
-            }));                    
-        },        
-    
+            }));
+        },
+
         showDeleteExpense: function(expense) {
             this.send('openModal', 'dialog', Ember.Object.create({
                 confirmAction: 'deleteExpense',
@@ -377,9 +377,9 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                 expenseToDelete: expense,
                 updateButtonAction: 'confirm',
                 updateButtonText: 'Ok'
-            }));   
+            }));
         },
-        
+
         showDeleteFamily: function(familyInfo) {
             this.send('openModal', 'dialog', Ember.Object.create({
                 confirmAction: 'deleteFamily',
@@ -388,14 +388,14 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                 familyToDelete: familyInfo,
                 updateButtonAction: 'confirm',
                 updateButtonText: 'Ok'
-            }));   
-            
+            }));
+
         },
 
         showDeleteImaging: function(imaging) {
             this.send('openModal', 'imaging.delete', imaging);
         },
-        
+
         showDeleteLab: function(lab) {
             this.send('openModal', 'labs.delete', lab);
         },
@@ -403,7 +403,7 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
         showDeleteMedication: function(medication) {
             this.send('openModal', 'medication.delete', medication);
         },
-        
+
         showDeletePhoto: function(photo) {
             this.send('openModal', 'dialog', Ember.Object.create({
                 confirmAction: 'deletePhoto',
@@ -419,32 +419,32 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
             visit.set('deleteFromPatient', true);
             this.send('openModal', 'visits.delete', visit);
         },
-    
+
         showEditExpense: function(model) {
             if (Ember.isEmpty(model)) {
                 model = this.get('store').createRecord('social-expense');
             }
-             this.send('openModal', 'patients.socialwork.expense', model);            
+             this.send('openModal', 'patients.socialwork.expense', model);
         },
-        
+
         showEditFamily: function(model) {
             if (Ember.isEmpty(model)) {
                 model = this.get('store').createRecord('family-info');
             }
             this.send('openModal', 'patients.socialwork.family-info', model);
         },
-        
+
         updateExpense: function(model) {
             var expenses = this.getWithDefault('expenses', []),
                 isNew = model.isNew;
-            if (isNew) {                
+            if (isNew) {
                 delete model.isNew;
-                expenses.addObject(model);                
+                expenses.addObject(model);
             }
             this.set('expenses', expenses);
             this.send('update', true);
             this.send('closeModal');
-        },        
+        },
 
         updateFamilyInfo: function(model) {
             var familyInfo = this.getWithDefault('familyInfo',[]),
@@ -461,17 +461,17 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
         updatePhoto: function(photo) {
             photo.save().then(function() {
                 this.send('closeModal');
-            }.bind(this));            
+            }.bind(this));
         },
-        
+
         visitDeleted: function(deletedVisit) {
             var visits = this.get('visits');
             visits.removeObject(deletedVisit);
             this.send('closeModal');
         }
-        
+
     },
-    
+
     _addChildObject: function(route) {
         this.transitionToRoute(route, 'new').then(function(newRoute) {
             newRoute.currentModel.setProperties( {
@@ -481,7 +481,7 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
             });
         }.bind(this));
     },
-    
+
     _getVisitCollection: function(name) {
         var returnList = [],
             visits = this.get('visits');
@@ -491,16 +491,16 @@ export default AbstractEditController.extend(BloodTypes, GenderList, ReturnTo, U
                     returnList.addObjects(items);
                     if (returnList.length > 0) {
                         returnList[0].set('first', true);
-                    }                    
+                    }
                 });
             });
         }
-        return returnList;        
+        return returnList;
     },
-    
+
     afterUpdate: function(record) {
         var message =  'The patient record for %@ has been saved.'.fmt(record.get('displayName'));
-        this.displayAlert('Patient Saved', message);        
+        this.displayAlert('Patient Saved', message);
     }
-    
+
 });
