+++ app/procedures/charge/controller.js
@@ -6,21 +6,21 @@ export default AbstractEditController.extend({
     cancelAction: 'closeModal',
     newPricingItem: false,
     requestingController: Ember.computed.alias('controllers.procedures/edit'),
-    pouchdb: Ember.inject.service(),
+    database: Ember.inject.service(),
     pricingList: Ember.computed.alias('controllers.procedures/edit.pricingList'),
-    
+
     updateCapability: 'add_charge',
 
     itemChanged: function() {
         var selectedItem = this.get('selectedItem');
         if (!Ember.isEmpty(selectedItem)) {
-            var pricingId = this.get('pouchdb').getEmberId(selectedItem._id);
+            var pricingId = this.get('database').getEmberId(selectedItem._id);
             this.store.find('pricing', pricingId).then(function(item) {
                 this.set('pricingItem', item);
             }.bind(this));
         }
     }.observes('selectedItem'),
-    
+
     pricingItemChanged: function() {
         var itemName = this.get('itemName'),
             pricingItem = this.get('pricingItem');
@@ -33,22 +33,22 @@ export default AbstractEditController.extend({
             this.set('newPricingItem', true);
         }
     }.observes('pricingItem'),
-    
+
     title: function() {
         var isNew = this.get('isNew');
         if (isNew) {
             return 'Add Charge Item';
         }
         return 'Edit Charge Item';
-	}.property('isNew'),    
-    
+	}.property('isNew'),
+
     beforeUpdate: function() {
         var isNew = this.get('isNew');
         if (isNew) {
-            this.set('newCharge', true);         
+            this.set('newCharge', true);
         }
         if (this.get('newPricingItem')) {
-            return new Ember.RSVP.Promise(function(resolve, reject) {                
+            return new Ember.RSVP.Promise(function(resolve, reject) {
                 var newPricing = this.store.createRecord('pricing', {
                     name: this.get('itemName'),
                     category: this.get('pricingCategory')
@@ -66,12 +66,12 @@ export default AbstractEditController.extend({
             return Ember.RSVP.Promise.resolve();
         }
     },
-    
+
     afterUpdate: function(record) {
-        if (this.get('newCharge')) {            
+        if (this.get('newCharge')) {
             this.get('requestingController').send('addCharge', record);
         } else {
             this.send('closeModal');
-        }        
+        }
     }
 });
